Add renderCharacter funcs
